refactor(infra): keep the fleet heartbeat resolver off the plugin SDK surface
heartbeat-summary.ts is wildcard re-exported by the plugin SDK, so exporting
resolveHeartbeatSummariesForAgents there widened the pinned public surface and
failed the SDK surface budget in CI. Move enrollment, the per-agent projection,
and the batched resolver into the internal heartbeat-summary-projection.ts;
the public module keeps exactly its previous exports and implements them
through the shared projection.
